Add files via upload
Problem Statement:  Write a program to perform following operations on Link List
a) Insertion 
	at the beginning
	at the end
	at the given location
in the sorted list
b) Deletion 
	of first node
	of last node
	of given item of node
	of given item from sorted list
c) Given a list, split it into two sublists — one for the front half, and one for the back half. 
If the number of elements is odd, the extra element should go in the front list. So 
FrontBackSplit() on the list {2, 3, 5, 7, 11} should yield the two lists {2, 3, 5} and {7, 11}.
d) Given two one-way-link lists A and B. Sort them independently and then Merge as list.
e) Two way link list insertion and Deletion
f) Header link list insertion and deletion